+++ CHANGELOG.md
@@ -2,4 +2,18 @@
 
 ### 0.0.0
 
-Initial release.
+- Initial release
+
+### 0.0.1
+
+- Added web support
+
+### 0.0.2
+
+- Added types, QuickNavBarType.static and QuickNavBarType.animate
+
+### 0.0.3
+
+- Improved documentation
+- Added example
+- Added new hoverEffect option for a hover effect